+++ couplet_composer/support/project_values.py
@@ -7,11 +7,13 @@ Composer acts on.
 """
 
 import json
+import logging
 import os
 
 from ..util.cache import cached
 
-from .file_paths import get_project_values_file_path
+from .file_paths import \
+    get_old_project_values_file_path, get_project_values_file_path
 
 from .project_names import get_ode_repository_name
 
@@ -26,13 +28,30 @@ def _get_project_values_file(source_root, in_tree_build):
 
     in_tree_build -- Whether the build files are created in tree.
     """
-    return os.path.join(source_root, get_project_values_file_path()) \
-        if in_tree_build \
-        else os.path.join(
-            source_root,
-            get_ode_repository_name(),
+    def _old_file():
+        return os.path.join(source_root, get_old_project_values_file_path()) \
+            if in_tree_build \
+            else os.path.join(
+                source_root,
+                get_ode_repository_name(),
+                get_old_project_values_file_path()
+            )
+    if os.path.isfile(_old_file()):
+        logging.warn(
+            "The file '%s' for providing project values is deprecated; use %s "
+            "instead",
+            get_old_project_values_file_path(),
             get_project_values_file_path()
         )
+        return _old_file()
+    else:
+        return os.path.join(source_root, get_project_values_file_path()) \
+            if in_tree_build \
+            else os.path.join(
+                source_root,
+                get_ode_repository_name(),
+                get_project_values_file_path()
+            )
 
 
 @cached
@@ -80,8 +99,15 @@ def get_project_version(source_root, in_tree_build):
         in_tree_build=in_tree_build
     )
     if project_values:
-        return None if "version" not in project_values \
-            else project_values["version"]
+        if "version" in project_values:
+            logging.warn(
+                "The use of the field 'version' for holding the shared "
+                "version number is deprecated; use 'shared_version' instead"
+            )
+            return project_values["version"]
+        else:
+            return None if "shared_version" not in project_values \
+                else project_values["shared_version"]
     else:
         return None